Buffer-Keepers/requirements.txt
pipe==2.0
eth-brownie==1.19.3
sentry-sdk==1.30.0
diskcache
requests[security]
redis
fastapi-cache2
retry
retry_requests
opentelemetry-distro
opentelemetry-exporter-otlp
web3==5.31.3
python-telegram-bot
numpy
sqlalchemy
psycopg2
pytz
pysha3